+++ web_service.go
@@ -6,7 +6,7 @@ import (
 	"reflect"
 	"sync"
 
-	"github.com/emicklei/go-restful/v3/log"
+	"github.com/yansong-easemob/go-restful/log"
 )
 
 // Copyright 2013 Ernest Micklei. All rights reserved.